+++ napari/layers/shapes/_mesh.py
