+++ napari/layers/shapes/_shapes_models/__init__.py
