+++ napari/layers/shapes/_shapes_utils.py
